manage state in posts component, posts refresh after submitting message
