+++ client/zestyreactnative/src/components/message/messageList.js
@@ -13,31 +13,12 @@ const ds = new ListView.DataSource({
   rowHasChanged: (r1, r2) => r1 !== r2
 });
 
-var dummyMessages = [
-  { user: 'Bronson', text: 'Nom nom nom', likes: -1 },
-  { user: 'Fifo', text: 'First in, first out. I mean: Woof!', likes: -1 }
-];
-
 class MessageList extends Component {
   constructor(props) {
     super(props);
-
-    this.state = {
-      dataSource: ds.cloneWithRows(dummyMessages),
-      errorMessage: ''
-    }
-
     this.renderMessage = this.renderMessage.bind(this);
   }
-
-  componentWillMount() {
-    getMessages()
-      .then((json) => {
-        return this.setState({ dataSource: ds.cloneWithRows(json) })
-      })
-      .catch((err) => this.setState({ errorMessage: err }) )
-  }
-
+  
   renderMessage({ User, text, likes }, index) {
     return (
       <Message
@@ -53,7 +34,7 @@ class MessageList extends Component {
     return (
       <View style={ [ styles.container ] }>
         <ListView
-          dataSource={ this.state.dataSource }
+          dataSource={ ds.cloneWithRows(this.props.data) }
           renderRow={ this.renderMessage }
         />
       </View>
