+++ client/zestyreactnative/src/components/timeline/posts.js
@@ -7,16 +7,38 @@ import {
 import MessageList from '../message/messageList';
 import Button from '../common/button';
 import { border } from '../../helpers/scaffolding';
+import { getMessages } from '../../helpers/api';
+
+var dummyMessages = [
+  { user: 'Bronson', text: 'Nom nom nom', likes: -1 },
+  { user: 'Fifo', text: 'First in, first out. I mean: Woof!', likes: -1 }
+];
 
 class Posts extends Component {
   constructor(props) {
     super(props);
-
     this.onPressCompose = this.onPressCompose.bind(this);
+    this.state = {
+      data: dummyMessages,
+      error: ''
+    }
+  }
+
+  componentWillMount() {
+    this.updateMessages();
   }
 
   onPressCompose() {
-    this.props.navigator.push({ name: 'messageComposer' });
+    this.props.navigator.push({name: 'messageComposer', callback: this.updateMessages.bind(this) });
+  }
+
+  updateMessages() {
+    console.log('getting new messages!');
+    getMessages()
+      .then((json) => {
+        this.setState({ data: json });
+      })
+      .catch((err) => this.setState({ error: err }) );
   }
 
   render() {
@@ -32,7 +54,7 @@ class Posts extends Component {
         </View>
 
         <View>
-          <MessageList />
+          <MessageList data={this.state.data} error={this.state.error} />
         </View>
       </View>
     );
